requirements.txt
torch
torchvision
timm
streamlit
pillow
matplotlib
numpy
tqdm
PyYAML
